requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
filelock==3.24.3
Flask==3.1.3
fonttools==4.61.1
fsspec==2026.2.0
gunicorn==25.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
packaging==26.0
pillow==12.1.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.10.0+cpu
torchaudio==2.10.0+cpu
torchvision==0.25.0+cpu
typing_extensions==4.15.0
Werkzeug==3.1.6
